fix(comp/singleDet): separated form
